requirements.txt
requests==2.31.0
pandas==2.1.4
matplotlib==3.8.0
fpdf2==2.7.4
numpy==1.24.3